moved chpl querier info to its own main function. removed functionality (for now) from endpointmanager main function.
